app/requirements/base.txt
amqp==2.1.1
anyjson==0.3.3
billiard==3.5.0.2
celery==4.0.0
data-importer==3.0.2
diff-match-patch==20121119
Django==1.8.17
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-crispy-forms==1.6.0
django-environ==0.4.1
django-filter==0.11.0
django-import-export==0.5.1
django-registration-redux==1.2
django-suit==0.2.21
django-taggit==0.22.0
django-tinymce==2.4.0
django-uuslug==1.1.8
djangorestframework==3.5.4
et-xmlfile==1.0.1
jdcal==1.3
kombu==4.0.0
Markdown==2.6.8
openpyxl==2.4.0
Pillow==3.2.0
pytz==2016.7
redis==2.10.5
requests==2.7.0
six==1.10.0
tablib==0.11.2
Unidecode==0.4.19
urllib3==1.20
vine==1.1.3
xlrd==1.0.0